refactor : move SearchCondition class package QUZ-47
