+++ search-service/search-domain/src/main/kotlin/com/grepp/quizy/search/domain/quiz/QuizSearchUseCase.kt
@@ -1,6 +1,5 @@
 package com.grepp.quizy.search.domain.quiz
 
-import com.grepp.quizy.search.domain.global.dto.SearchCondition
 import com.grepp.quizy.search.domain.global.dto.Slice
 import com.grepp.quizy.search.domain.user.UserId
