+++ search-service/search-infra/src/main/kotlin/com/grepp/quizy/search/infra/quiz/repository/QuizSearchRepositoryAdapter.kt
@@ -1,7 +1,7 @@
 package com.grepp.quizy.search.infra.quiz.repository
 
 import com.grepp.quizy.search.domain.global.dto.Slice
-import com.grepp.quizy.search.domain.global.dto.SearchCondition
+import com.grepp.quizy.search.domain.quiz.SearchCondition
 import com.grepp.quizy.search.domain.quiz.*
 import com.grepp.quizy.search.domain.user.UserId
 import com.grepp.quizy.search.infra.quiz.document.QuizDomainFactory
